Fix paths to handler
